requirements.txt
pytouch
torchvision
tensorboard
tensorboardX
matplotlib
python-mnist
pyyaml
sklearn
